refactor(jest): detect jest version for testing
add a new abstraction layer to stencil's jest testing
configuration/running such that the version of jest is detected when
jest is run. based on the version that is detected, defer to the
correct piece of versioned infrastructure to setup and run tests.

in https://github.com/ionic-team/stencil/pull/4847, the directory
structure of the testing submodule was updated to look something like:
```
├── src/testing/jest/
│   └── jest-27-and-under
│       ├── jest-config.ts
│       ├── jest-environment.ts
│       ├── jest-preprocessor.ts
│       ├── jest-runner.ts
│       ├── jest-screenshot.ts
│       ├── jest-serializer.ts
│       ├── jest-setup-test-framework.ts
│       ├── node_modules/
│       ├── package.json
│       └── test
└── package.json
```

this commit introduces the following infrastructue to dispatch to
jest-27-and-under at runtime:

```diff
  ├── src/testing/jest/
  │   └── jest-27-and-under
  │       ├── jest-config.ts
  │       ├── jest-environment.ts
+ │       ├── jest-facade.ts                  <- Implements base class in `jest-adapter.ts`
  │       ├── jest-preprocessor.ts
  │       ├── jest-runner.ts
  │       ├── jest-screenshot.ts
  │       ├── jest-serializer.ts
  │       ├── jest-setup-test-framework.ts
  │       ├── node_modules/
  │       ├── package.json
  │       └── test
+ ├── jest-adapter.ts                         <- Base class all `jest-facade` implementations should follow
+ ├── jest-apis.ts                            <- Interactions with Jest prior to knowing the version
+ ├── jest-stencil-connector.ts               <- Routes Jest logic through the correct `jest-facade` implementation
  └── package.json
```

Where:
- `jest-adapter` - a file containing a base class to define the API by
which version-specific implementations can be loaded. a jest
v27-specific implementation is included as well in `jest-27-and-under/`
- `jest-stencil-connector` - performs the detection of the current
version of jest - dispatching to the correct infrastructure via
`jest-facade` implementations
- `jest-apis` - a file containing typings for cases where the jest
typings are ambiguous, particularly for when we do not know what the
current version of jest is (yet)

the end result of this commit should be nothing to the end user (for
now). this work shall be used to help support jest v28+ in the near
future (dispatching the the correct version)